Fix: Add confirmation before deleting data

+++ modules/document/views/default/index.php
@@ -39,7 +39,7 @@ $this->parameters = [
                                 </td>
                                 <td class="text-right">
                                     <a class="btn btn-secondary btn-sm" href="index.php?r=document/default/update&id=<?= $model->id ?>"><i class="fas fa-pencil-alt"></i></a>
-                                    <a class="btn btn-secondary btn-sm" href="index.php?r=document/default/delete&id=<?= $model->id ?>"><i class="far fa-trash-alt"></i></a>                                    
+                                    <a class="btn btn-secondary btn-sm confirmation-modal" href="index.php?r=document/default/delete&id=<?= $model->id ?>"><i class="far fa-trash-alt"></i></a>                                    
                                 </td>
                             </tr>
                         <?php endforeach; ?>
